requirements.txt

streamlit
langchain==0.3.27
langchain-core
langchain-community
langchain-groq
python-dotenv
sentence-transformers
faiss-cpu
PyPDF2

